update widget code
